requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
djongo==1.3.6
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4
